inference/requirements.txt
torch
torchvision
pillow
numpy
uvicorn
fastapi
torchsummary